--- Assets/Scripts/MoveAction.cs
@@ -1,84 +1,0 @@
-using System.Collections.Generic;
-using UnityEngine;
-
-public class MoveAction : MonoBehaviour
-{
-    [SerializeField] private Unit unit;
-    [SerializeField] private int maxMoveDistance = 2;
-
-    private Vector3 targetPosition;
-
-    public bool IsWalking { get; private set; }
-
-    private void Awake()
-    {
-        targetPosition = transform.position;
-    }
-
-    private void Update()
-    {
-        float stoppingDistance = 0.1f;
-        if (Vector3.Distance(targetPosition, transform.position) > stoppingDistance)
-        {
-            Vector3 moveDirection = (targetPosition - transform.position).normalized;
-            float moveSpeed = 4f;
-            transform.position += moveDirection * moveSpeed * Time.deltaTime;
-            IsWalking = true;
-
-            float rotateSpeed = 10f;
-            transform.forward = Vector3.Lerp(transform.forward, moveDirection, rotateSpeed * Time.deltaTime);
-        }
-        else
-        {
-            IsWalking = false;
-        }
-    }
-
-    public bool IsValidActionGridPosition(GridPosition gridPosition)
-    {
-        List<GridPosition> validActionGridPositionList = GetValidActionGridPositionList();
-        return validActionGridPositionList.Contains(gridPosition);
-    }
-
-    public List<GridPosition> GetValidActionGridPositionList()
-    {
-        List<GridPosition> validActionGridPositionList = new List<GridPosition>();
-
-        GridPosition unitGridPosition = unit.GetGridPosition();
-        for (int x = -maxMoveDistance; x <= maxMoveDistance; x++)
-        {
-            for (int z = -maxMoveDistance; z <= maxMoveDistance; z++)
-            {
-                GridPosition offsetGridPosition = new GridPosition(x, z);
-                GridPosition testGridPosition = offsetGridPosition + unitGridPosition;
-
-                if(!LevelGrid.Instance.IsValidGridPosition(testGridPosition))
-                {
-                    //Not valid GridPosition
-                    continue;
-                }
-
-                if(testGridPosition == unitGridPosition)
-                {
-                    //Unit is already located in that GridPosition
-                    continue;
-                }
-
-                if (LevelGrid.Instance.HasAnyUnitOnGridPosition(testGridPosition))
-                {
-                    //GridPosition is already occupied by another unit
-                    continue;
-                }
-
-                validActionGridPositionList.Add(testGridPosition);
-            }
-        }
-
-        return validActionGridPositionList;
-    }
-
-    public void Move(GridPosition gridPosition)
-    {
-        targetPosition = LevelGrid.Instance.GetWorldPosition(gridPosition);
-    }
-}
